added ability to test configs
Currently, the config test functions are in config.js,
but I think they will be moved when my testing tools.

`emptyCheck` does what it says in the name;
it confirms that a variable is defined.

`fileExists` if confirms the variable it is check is set,
then it confirms the file exists.
`BASE_PATH` is not used atm,
but was used as a stand in config option.